整理 task 代码

+++ idc-app/src/main/java/com/iwellmass/idc/app/service/WorkflowService.java
@@ -1,6 +1,7 @@
 package com.iwellmass.idc.app.service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import javax.inject.Inject;
 
@@ -8,6 +9,8 @@ import com.iwellmass.idc.model.TaskKey;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.google.common.base.Charsets;
+import com.google.common.hash.Hashing;
 import com.iwellmass.idc.IDCUtils;
 import com.iwellmass.idc.app.repo.WorkflowEdgeRepository;
 import com.iwellmass.idc.app.repo.WorkflowRepository;
@@ -25,17 +28,29 @@ public class WorkflowService {
 
 
     @Transactional
-    public Workflow saveWorkflow(Workflow workflow) {
+    public boolean saveWorkflow(Workflow workflow) {
+    	
     	List<WorkflowEdge> edges = IDCUtils.parseWorkflowEdge(workflow.getGraph());
     	
-    	edges.forEach(we -> {
-    		we.setParentTaskKey(new TaskKey(workflow.getTaskId(),workflow.getTaskGroup()));
-    	});
+    	List<String> ks = edges.stream().sorted().map(e -> {
+    		return e.getTaskKey().toString() + e.getSrcTaskKey().toString();
+    	}).collect(Collectors.toList());
     	
-    	workflowEdgeRepository.deleteByParentTaskIdAndParentTaskGroup(workflow.getTaskId(),workflow.getTaskGroup());
-    	workflowEdgeRepository.save(edges);
-    	workflowRepository.save(workflow);
-    	return workflow;
+    	String autoId = Hashing.md5().hashString(String.join(",", ks), Charsets.UTF_8).toString();
+
+    	Workflow check = workflowRepository.findOne(autoId);
+    	if (check == null) {
+    		workflow.setWorkflowId(autoId);
+    		edges.forEach(we -> {
+    			we.setParentTaskKey(new TaskKey(workflow.getTaskId(),workflow.getTaskGroup()));
+    		});
+    		workflowRepository.save(workflow);
+    		// 刷新 edges
+    		workflowEdgeRepository.deleteByParentTaskIdAndParentTaskGroup(workflow.getTaskId(),workflow.getTaskGroup());
+    		workflowEdgeRepository.save(edges);
+    		return true;
+    	}
+    	return false;
     }
 
 	public Workflow getWorkflow(String workflowId) {
